Fixing deadlock conditions involving studio join
Removing VsQuickView from VsAuth since there seems to be
no good reason for it to have visibility into that, and
it breaks the pattern of VirtualStudio owning that
exclusively.

Removed the authSucceeded and authFailed signals
from VirtualStudio because: (a) these were duplicates
from VsAuth, (b) nothing used authFailed, (c) there
was only a single qml connection really using
authSucceeded and it was easy to wire that up directly
instead, since auth is a context variable, (d) having
these in both places and meaning different things was
very confusing and made it harder to grok and maintain

Replaced the firstLoad parameter in VirtualStudio::serverList
with a m_firstRefresh class member, which is a lot easier to
track and understand.

Changing use of m_refreshMutex to be a guard specifically
for any access to m_servers or m_refreshInProgress. Fixing
a few cases where these were not handled properly.

Fixing various compiler warnings in other source code

+++ src/AudioInterface.cpp
@@ -77,7 +77,7 @@ AudioInterface::AudioInterface(QVarLengthArray<int> InputChans,
 #ifndef WAIR
     // cc
     // Initialize and assign memory for ProcessPlugins Buffers
-    int monitorChans = std::min<size_t>(mInputChans.size(), mOutputChans.size());
+    int monitorChans = int(std::min<size_t>(mInputChans.size(), mOutputChans.size()));
     mInProcessBuffer.resize(mInputChans.size());
     mOutProcessBuffer.resize(mOutputChans.size());
     mMonProcessBuffer.resize(monitorChans);
@@ -187,8 +187,8 @@ void AudioInterface::setup(bool /*verbose*/)
     // Allocate buffer memory to read and write
     mSizeInBytesPerChannel = getSizeInBytesPerChannel();
 
-    int size_audio_input  = mSizeInBytesPerChannel * nChansIn;
-    int size_audio_output = mSizeInBytesPerChannel * nChansOut;
+    int size_audio_input  = int(mSizeInBytesPerChannel * nChansIn);
+    int size_audio_output = int(mSizeInBytesPerChannel * nChansOut);
 #ifdef WAIR               // WAIR
     if (mNumNetRevChans)  // else don't change sizes
     {
